requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras
#
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==5.0.6
    # via
    #   django-bootstrap5
    #   djangorestframework
    #   drf-spectacular
    #   shurl_django (pyproject.toml)
django-bootstrap5==24.2
    # via shurl_django (pyproject.toml)
django-environ==0.11.2
    # via shurl_django (pyproject.toml)
djangorestframework==3.15.2
    # via drf-spectacular
drf-spectacular==0.27.2
    # via shurl_django (pyproject.toml)
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.22.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
psycopg==3.1.19
    # via shurl_django (pyproject.toml)
psycopg-binary==3.1.19
    # via psycopg
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.0
    # via django
typing-extensions==4.12.2
    # via psycopg
tzdata==2024.1
    # via
    #   django
    #   psycopg
uritemplate==4.1.1
    # via drf-spectacular
